# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farah API Documentation",
        "description": "توثيق واجهات برمجة التطبيقات (APIs) لمنصة فرح لتطبيق Flutter",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://farahbackend-production.up.railway.app",
            "description": "سيرفر الإنتاج (Railway Production)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "السيرفر المحلي (Local Server)"
        }
    ],
    "paths": {
...
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "جلب قائمة عناوين المستخدم الحالي",
                "operationId": "4a95f25c1e13530115fd3fd6961b92de",
                "responses": {
                    "200": {
                        "description": "تم استرجاع العناوين بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "icon": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Locations retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "إضافة عنوان جديد للمستخدم الحالي",
                "operationId": "2e940d1dad83fe7c7fb5a308a7f6d84a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "المنزل"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "شارع الجلاء، غزة"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 31.5
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.4667
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء العنوان بنجاح"
                    },
                    "422": {
                        "description": "خطأ في البيانات المدخلة"
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/locations/{location}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "عرض تفاصيل عنوان محدد",
                "operationId": "f167a429e4fe99e390b7f6e5fb2a0b0e",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "معرّف العنوان",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم استرجاع العنوان بنجاح"
                    },
                    "403": {
                        "description": "العنوان غير تابع للمستخدم الحالي"
                    },
                    "404": {
                        "description": "العنوان غير موجود"
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "تعديل عنوان محدد",
                "operationId": "879c97824fb3fbada5295f184e5385b9",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "معرّف العنوان",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "العمل"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "شارع الرمال، غزة"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 31.51
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث العنوان بنجاح"
                    },
                    "403": {
                        "description": "العنوان غير تابع للمستخدم الحالي"
                    },
                    "422": {
                        "description": "خطأ في البيانات المدخلة"
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "حذف عنوان محدد",
                "operationId": "b087bde5ca5fc36bf09597126337778c",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "معرّف العنوان",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم حذف العنوان بنجاح"
                    },
                    "403": {
                        "description": "العنوان غير تابع للمستخدم الحالي"
                    },
                    "404": {
                        "description": "العنوان غير موجود"
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/locations/{location}/primary": {
            "patch": {
                "tags": [
                    "Locations"
                ],
                "summary": "تعيين العنوان الأساسي",
                "operationId": "2b33624e9dc6008795ea47d110765121",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تعيين العنوان الأساسي"
                    },
                    "403": {
                        "description": "العنوان لا يخص المستخدم"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل التوكن بهذا الشكل: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}